Changed background image file
Linked to smaller version of file. Also put in some media query commands
that should make use of a smaller file for tablets and phones. Currently
I have a different image so that I can deploy and test, but I might
switch it back to the same image as the desktop version. May not matter
since you can't see much on phones and tablets.